requirements.txt
pdfplumber>=0.11,<0.12
google-api-python-client>=2.0,<3.0
google-auth>=2.0,<3.0
google-auth-oauthlib>=1.0,<2.0
google-auth-httplib2>=0.2,<1.0
python-dotenv>=1.0,<2.0
